+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- fix opencode quick-answer
+
 ## [0.3.1] - 2026-04-02
 
 - `tj init`: interactive setup wizard to create initial config file